Renamed PrefixHandler.py, removed !d commands, replaced by !roll
